Flash messages and User model
Installed the following npm packages. (connect-flash, passport, passport-local, passport-local-mongoose)
Reinstalled version 5.11.12 of mongoose. (I accidentally updated it)
Added a dismissable flash message to appear when adding, editing, or deleting a campground with a custom message.
Added a dismissable flash message to appear when adding and deleting a review with a custom message.
Added a flash message and a redirect when visiting a no longer existing campground.
Added a User model and a temporary test route at /fakeuser to test creating user accounts with passport.