Added Register/Login/Logout functionality
Added a users route.
Renamed our route variables in the app.js file to make clearer what they are.
Added a register page that flashes a welcome message when a user registers.
Added a login page that flashes a welcome back message when a user logs in.
Added a 'middleware' folder.
User must be logged in order to add, edit or delete a campground. If they are not they will be redirected to login.
If a user is redirected to log in, if they do so they will be redirected back to the page they were on.
Added a /logout page that logs out the user.
Added more links to the navbar. Navbar links will change based on if a user is logged in.