Customizing Cluster Map Pop Up Markers
Created a properties.popUpMarkup virtual on the campgroundSchema for MapBox.
Allowed JSON virtuals on the campgroundSchema.
Customized the Cluster Map Pop Up to show the title and description that links to that campground.
Added a mouse pointer when hovering over a campground on the Cluster Map.
Updated some variables from Var to Const.